chore: trending API 재요청 막기
