+++ src/index.tsx
@@ -1,5 +1,11 @@
 import { createRoot } from 'react-dom/client';
 import App from './App';
 
+import { GifContextProvider } from './context/Gif';
+
 const root = createRoot(document.getElementById('app')!);
-root.render(<App />);
+root.render(
+  <GifContextProvider>
+    <App />
+  </GifContextProvider>
+);
