+++ src/pages/Search/components/SearchResult/SearchResult.tsx
@@ -3,7 +3,7 @@ import { GifImageModel } from '../../../../models/image/gifImage';
 import ResultTitle from '../ResultTitle/ResultTitle';
 import GifItem from '../GifItem/GifItem';
 
-import { SearchStatus, SEARCH_STATUS } from '../../hooks/useGifSearch';
+import { SearchStatus, SEARCH_STATUS } from '../../../../context/Gif';
 
 import styles from './SearchResult.module.css';
