+++ src/pages/Search/hooks/useGifSearch.tsx
@@ -1,23 +1,16 @@
-import { ChangeEvent, useEffect, useState } from 'react';
+import { ChangeEvent, useContext, useEffect, useState } from 'react';
 
 import { gifAPIService } from '../../../apis/gifAPIService';
+import GifContext from '../../../context/Gif';
 import { GifImageModel } from '../../../models/image/gifImage';
+import { SEARCH_STATUS } from '../../../context/Gif';
 
 const DEFAULT_PAGE_INDEX = 0;
 
-export const SEARCH_STATUS = {
-  BEFORE_SEARCH: 'BEFORE_SEARCH',
-  LOADING: 'LOADING',
-  FOUND: 'FOUND',
-  NO_RESULT: 'NO_RESULT'
-} as const;
-
-export type SearchStatus = typeof SEARCH_STATUS[keyof typeof SEARCH_STATUS];
-
 const useGifSearch = () => {
-  const [status, setStatus] = useState<SearchStatus>(SEARCH_STATUS.BEFORE_SEARCH);
+  const { status, setStatus, gifList, setGifList } = useContext(GifContext);
+
   const [currentPageIndex, setCurrentPageIndex] = useState(DEFAULT_PAGE_INDEX);
-  const [gifList, setGifList] = useState<GifImageModel[]>([]);
   const [searchKeyword, setSearchKeyword] = useState('');
 
   const updateSearchKeyword = (e: ChangeEvent<HTMLInputElement>) => {
@@ -60,11 +53,10 @@ const useGifSearch = () => {
         const gifs: GifImageModel[] = await gifAPIService.getTrending();
 
         setGifList(gifs);
+        setStatus(SEARCH_STATUS.FOUND);
       }
     };
     fetch();
-
-    return () => setStatus(SEARCH_STATUS.LOADING);
   }, []);
 
   return {
